// src/stores/user/interfaces/userData.ts
export interface UserData {
  id: number;
  token: string;
  name: string;
  email: string;
  email_verified_at: string;
  phone: string;
  role: string;
  parent_phone: string;
  study_level: string;
  department: string;
  is_blocked: number;
  is_active: number;
  created_at: string;
  updated_at: string;
}